[#36] refactor: 매직 넘버 상수 처리
